Azure-Container-Apps/AI-Model-YoloV5-Dapr-Queue/requirements.txt
requests
black

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.41.0

tensorboard>=2.4.1

seaborn>=0.11.0
pandas

thop  # FLOPs computation

dapr 
dapr-ext-grpc